Reorg functions in files
